+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubValidations (development version)
+
 # hubValidations 2.0.0
 
 ## Breaking Changes